Refactored bar chart and pie chart
- Fixed compile errors
- Refactored bar chart and pie chart

Not tested because of the error "requirement failed" and I didn't
understood the cause of this error.

Please note: Since the origin is no longer at (0,0) in the svg the bar
chart and pie chart probably will be displaced. Also the grid of a line
chart is not correctly adjusted.